requirements.txt
fastapi==0.136.0
uvicorn[standard]==0.44.0
python-multipart==0.0.20
pydantic-settings==2.13.1
yt-dlp==2026.3.17
static-ffmpeg==3.0
groq==1.1.2
google-genai==1.73.1
assemblyai==0.61.1
deepgram-sdk==6.1.1
httpx==0.28.1
